test(unity): split daemon logger coverage
Move daemon exception logging coverage out of UnityLogCollectorTests into a dedicated DaemonLoggerTests fixture so each test class stays aligned with a single runtime concern.
