+++ src/Ucli.Unity/Assets/Tests/MackySoft.Ucli.Unity.Tests/Editor/DaemonLoggerTests.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 7e00e45f9cf6a4097bea166e2e14af42
